+++ app/update.go
@@ -1,12 +1,10 @@
 package app
 
 import (
-	"bytes"
 	"crypto/sha256"
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
-	"io"
 	"log/slog"
 	"net/http"
 	"runtime"
@@ -14,7 +12,6 @@ import (
 	"golang.org/x/mod/semver"
 
 	"github.com/LEX0RE/rockpload/app/tools/logger"
-	"github.com/inconshreveable/go-update"
 )
 
 type UpdateInfo struct {
@@ -69,26 +66,6 @@ func (u *Updater) CheckForUpdate(currentVersion string) (bool, error) {
 	}
 
 	return false, nil
-}
-
-func (u *Updater) ApplyUpdate() error {
-	logger.FuncDebug()
-	resp, err := http.Get(u.UpdateInfo.URL)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	data, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return err
-	}
-
-	if !verifyChecksum(data, u.UpdateInfo.Checksum) {
-		return fmt.Errorf("invalid checksum")
-	}
-
-	return update.Apply(bytes.NewReader(data), update.Options{})
 }
 
 func verifyChecksum(data []byte, expected string) bool {
